fix(security): keep service-role client out of the browser bundle (vault)
updateIntegrationMetadata (added in the vault work) did `await import(
"@carbon/auth/client.server")` — the service-role, RLS-bypassing client — inside
settings.service.ts, which the ~/modules/settings barrel (export *) pulls into the
CLIENT bundle. Vite's server-code guard was correctly refusing to ship it; it would
also break the production build.

Move the function to settings.server.ts (server-only, not in the barrel), which
already imports getCarbonServiceRole statically — no dynamic import. It has no other
callers; it was only reachable as the MCP tool settings_updateIntegrationMetadata,
which regenerating tool-metadata.json now drops (1442->1441) — writing integration
secrets via MCP is not a surface we want anyway.
